+++ test/e2e/flask/btc/create-btc-account.spec.ts
@@ -12,6 +12,7 @@ import { ACCOUNT_TYPE } from '../../constants';
 import { withBtcAccountSnap } from './common-btc';
 
 describe('Create BTC Account', function (this: Suite) {
+  // eslint-disable-next-line mocha/no-skipped-tests
   it.skip('create BTC account from the menu', async function () {
     await withBtcAccountSnap(
       { title: this.test?.fullTitle() },
@@ -23,6 +24,7 @@ describe('Create BTC Account', function (this: Suite) {
     );
   });
 
+  // eslint-disable-next-line mocha/no-skipped-tests
   it.skip('cannot create multiple BTC accounts', async function () {
     await withBtcAccountSnap(
       { title: this.test?.fullTitle() },
@@ -46,6 +48,7 @@ describe('Create BTC Account', function (this: Suite) {
     );
   });
 
+  // eslint-disable-next-line mocha/no-skipped-tests
   it.skip('can cancel the removal of BTC account', async function () {
     await withBtcAccountSnap(
       { title: this.test?.fullTitle() },
@@ -70,6 +73,7 @@ describe('Create BTC Account', function (this: Suite) {
     );
   });
 
+  // eslint-disable-next-line mocha/no-skipped-tests
   it.skip('can recreate BTC account after deleting it', async function () {
     await withBtcAccountSnap(
       { title: this.test?.fullTitle() },
@@ -115,6 +119,7 @@ describe('Create BTC Account', function (this: Suite) {
     );
   });
 
+  // eslint-disable-next-line mocha/no-skipped-tests
   it.skip('can recreate BTC account after restoring wallet with SRP', async function () {
     await withBtcAccountSnap(
       { title: this.test?.fullTitle() },
