+++ test/e2e/tests/confirmations/transactions/native-send-redesign.spec.ts
@@ -17,7 +17,7 @@ import { TestSuiteArguments } from './shared';
 const TOKEN_RECIPIENT_ADDRESS = '0x2f318C334780961FB129D2a6c30D0763d9a5C970';
 
 describe('Confirmation Redesign Native Send', function () {
-  describe('Wallet initiated', async function () {
+  describe('Wallet initiated', function () {
     it('Sends a type 0 transaction (Legacy)', async function () {
       await withTransactionEnvelopeTypeFixtures(
         this.test?.fullTitle(),
@@ -39,7 +39,7 @@ describe('Confirmation Redesign Native Send', function () {
     });
   });
 
-  describe('dApp initiated', async function () {
+  describe('dApp initiated', function () {
     it('Sends a type 0 transaction (Legacy)', async function () {
       await withTransactionEnvelopeTypeFixtures(
         this.test?.fullTitle(),
