+++ test/e2e/tests/content-security-policy/content-security-policy.spec.ts
@@ -7,6 +7,7 @@ import { loginWithBalanceValidation } from '../../page-objects/flows/login.flow'
 
 describe('Content-Security-Policy', function (this: Suite) {
   // TODO: Re-enable this after fixing the CSP override feature. See #31094
+  // eslint-disable-next-line mocha/no-skipped-tests
   it.skip('opening a restricted website should still load the extension', async function () {
     await withFixtures(
       {
